requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
tenacity>=8.2.0
diskcache>=5.6.0
python-dotenv>=1.0.0
pygments>=2.17.0
tiktoken>=0.7.0
